local authorities downloader ported to new strategy-based caching mechanism, and tested from console (not automated tests yet)

+++ postcodeinfo/apps/postcode_api/caches/filesystem_cache.py
@@ -13,18 +13,16 @@ from time import mktime
 
 from postcode_api.caches.cache import Cache
 
+log = logging.getLogger(__name__)
+
 
 class FilesystemCache(Cache):
 
     def __init__(self, *args, **kwargs):
         self._bucket = None
         self.dir = kwargs.pop('dir', '/tmp/')
-
-    def key_for(self, object_identifier):
-        """
-        Assumption: we'll be given valid filenames (?)
-        """
-        return object_identifier
+        if not os.path.exists(self.dir):
+            os.makedirs(self.dir)
 
     def _full_path(self, cache_key):
         return os.path.join(self.dir, cache_key)
@@ -33,13 +31,23 @@ class FilesystemCache(Cache):
         return os.path.isfile(self._full_path(cache_key))
 
     def get(self, cache_key, dest_filename):
-        return shutil.copy2(self._full_path(cache_key), dest_filename)
+        if self.has(cache_key):
+            return shutil.copy2(self._full_path(cache_key), dest_filename)
 
-    def put(self, filename):
-        shutil.copy2(filename, self._full_path(self.key_for(filename)))
+    def put(self, cache_key, filename):
+        dest_path = self._full_path(cache_key)
+        log.info(
+            "putting {filename} to filesystem cache with key {key}, full path = {path}"
+            .format(filename=filename, key=cache_key, path=dest_path))
+        if filename == dest_path:
+            log.error("{filename} and {dest_path} are the same - nothing to do!".format(
+                filename=filename, dest_path=dest_path))
+        else:
+            shutil.copy2(filename, dest_path)
 
     def delete(self, cache_key):
-        return os.remove(self._full_path(cache_key))
+        if self.has(cache_key):
+            return os.remove(self._full_path(cache_key))
 
     def last_modified(self, cache_key):
         mtime = getmtime(self._full_path(cache_key))
